requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
Pillow==8.4.0
